Update report_template.tex
Added Synopsis to the table of contents, also made the synopsis page 3 of the report rather than page 1 so as to meet the specifications in the style guide  demo page (section 3.2.1)

I also added an example of how to use the \nomenclature command, and added the lines necessary to display the nomenclature section in the report. The specific command I added makes use of a unit property, for which a new command was made in the preamble. This added command is as suggested within the documentation for the package, I did not come up with the command myself.

As is standard with makeindex and also with nomencl, before the nomenclature is displayed, the nomeclature file %.nls must first be generated. I have added this comment within the preamble
 with a basic command that can be saved within the IDE of choice as a custom command.