Biblatex custom config file and code cleanup
All customisation to the default author-year biblatex style has been
moved to the configuration file "biblatex.cfg". The name of the file
must remain as it is and it must have the same path as the tex file
being created. This is because biblatex calls the config file after
including the specified style and before compilation of the
bibliography.

The config file contains an edited version of the patent driver, as well
as customisation of macros and bibliography settings. Comments can be
included if explanation of the customisation will add value?

As far as cleanup is concerned, predefined customisation tools exist in
biblatex (custom versions of \renewcommand etc.) and were implemented
where possible.