New biblatex configurations
Added several new edits to the configuration. Most edits relate to
punctuation or notations of articles, books and incollection drivers.

Modified the renewal of the "in:" macro to only be eliminated from the
journal driver as it has correct usage in incollection and inproceeding
drivers.

Fixed fixed the page number presentation of citations and references
(removing p. and pp.).

Fixed the sorting and presentation of editors in articles and
incollections, possibly in other drivers (untested).